Current Unfollow Class
This is my unfollow class that I made the sign in and the movement from "home" to "profile" works just fine, but the program keeps failing when it tries to select the "following" link I've tried experimenting with different web element identifiers and even setting the program to wait until the element is selectable. Yet it keeps failing at this crucial point. 
Also you might want to set up your chromedriver.exe environment for your computer and set up an Instagram account if you want to run this.